[feat] Organise views folder
