+++ src/views/DcOpenOrders.tsx
@@ -1,12 +1,12 @@
 import { Column } from "primereact/column";
 import { FC, useEffect, useState } from "react";
-import Card from "./common/Card";
-import CustomTable from "./common/customtable";
-import Loader from "./loader";
+import Card from "../components/common/Card";
+import CustomTable from "../components/common/customtable";
+import Loader from "../components/loader";
 import { DCOpenOrders, orderData } from "../@types/pages/dcOpenOrders";
 import { URL_DC_OPEN_ORDERS } from "../constants/apiConstants";
 import { fetchData } from "../utils/fetchUtil";
-import { getTableHeaders } from "./utils/Utils";
+import { getTableHeaders } from "../components/utils/Utils";
 import { PAGE_TITLES } from "../constants/appConstants";
 
 const DcOpenOrders: FC = () => {
@@ -41,7 +41,7 @@ const DcOpenOrders: FC = () => {
       ) : (
         <div>
           <h3 className="sm:text-lg text-gray-200 font-bold m-3 ml-0 font-helvetica">
-              {PAGE_TITLES.DC_OPEN_ORDERS}
+            {PAGE_TITLES.DC_OPEN_ORDERS}
           </h3>
           <div className="hidden sm:block rounded-md">
             {tableData?.length > 0 && (
